code pushed to tano sabers

+++ src/saber/Grinliz_Util.cpp
@@ -6,7 +6,7 @@ static int16_t gSinTable[256] = { 0 };
 uint8_t lerpU8(uint8_t a, uint8_t b, uint8_t t) 
 {
     int32_t r = int32_t(a) + (int32_t(b) - int32_t(a)) * int32_t(t) / 255;
-    return uint8_t(clamp(r, 0, 255));
+    return uint8_t(clamp(r, int32_t(0), int32_t(255)));
 }
 
 bool TestUtil()
